+++ lib/prefs.dart
@@ -30,7 +30,7 @@ import 'package:shared_preferences/shared_preferences.dart'
 
 /// The App's Preferences.
 class Prefs {
-  static Future<SharedPreferences> get instance async =>
+  static Future<SharedPreferences> get _instance async =>
       _prefs ??= await SharedPreferences.getInstance();
   static SharedPreferences _prefs;
 
@@ -42,7 +42,7 @@ class Prefs {
   /// Initialize the SharedPreferences object in the State object's iniState() function.
   static Future<SharedPreferences> init() async {
     _initCalled = true;
-    _prefsInstance = await instance;
+    _prefsInstance = await _instance;
     return _prefsInstance;
   }
 
@@ -65,7 +65,7 @@ class Prefs {
   static Future<Set<String>> getKeysF() async {
     Set<String> value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.getKeys() ?? Set();
     } else {
       value = getKeys();
@@ -86,7 +86,7 @@ class Prefs {
   static Future<dynamic> getF(String key) async {
     dynamic value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.get(key);
     } else {
       value = get(key);
@@ -106,7 +106,7 @@ class Prefs {
   static Future<bool> getBoolF(String key, [bool defValue]) async {
     bool value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.getBool(key) ?? defValue ?? false;
     } else {
       value = getBool(key);
@@ -126,7 +126,7 @@ class Prefs {
   static Future<int> getIntF(String key, [int defValue]) async {
     int value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.getInt(key) ?? defValue ?? 0;
     } else {
       value = getInt(key);
@@ -146,7 +146,7 @@ class Prefs {
   static Future<double> getDoubleF(String key, [double defValue]) async {
     double value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.getDouble(key) ?? defValue ?? 0.0;
     } else {
       value = getDouble(key);
@@ -166,7 +166,7 @@ class Prefs {
   static Future<String> getStringF(String key, [String defValue]) async {
     String value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.getString(key) ?? defValue ?? "";
     } else {
       value = getString(key);
@@ -187,7 +187,7 @@ class Prefs {
       [List<String> defValue]) async {
     List<String> value;
     if (_prefsInstance == null) {
-      var prefs = await instance;
+      var prefs = await _instance;
       value = prefs?.getStringList(key) ?? defValue ?? [""];
     } else {
       value = getStringList(key);
@@ -198,14 +198,14 @@ class Prefs {
   /// Saves a boolean [value] to persistent storage in the background.
   /// If [value] is null, this is equivalent to calling [remove()] on the [key].
   static Future<bool> setBool(String key, bool value) async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.setBool(key, value) ?? Future.value(false);
   }
 
   /// Saves an integer [value] to persistent storage in the background.
   /// If [value] is null, this is equivalent to calling [remove()] on the [key].
   static Future<bool> setInt(String key, int value) async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.setInt(key, value) ?? Future.value(false);
   }
 
@@ -213,33 +213,33 @@ class Prefs {
   /// Android doesn't support storing doubles, so it will be stored as a float.
   /// If [value] is null, this is equivalent to calling [remove()] on the [key].
   static Future<bool> setDouble(String key, double value) async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.setDouble(key, value) ?? Future.value(false);
   }
 
   /// Saves a string [value] to persistent storage in the background.
   /// If [value] is null, this is equivalent to calling [remove()] on the [key].
   static Future<bool> setString(String key, String value) async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.setString(key, value) ?? Future.value(false);
   }
 
   /// Saves a list of strings [value] to persistent storage in the background.
   /// If [value] is null, this is equivalent to calling [remove()] on the [key].
   static Future<bool> setStringList(String key, List<String> value) async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.setStringList(key, value) ?? Future.value(false);
   }
 
   /// Removes an entry from persistent storage.
   static Future<bool> remove(String key) async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.remove(key) ?? Future.value(false);
   }
 
   /// Completes with true once the user preferences for the app has been cleared.
   static Future<bool> clear() async {
-    var prefs = await instance;
+    var prefs = await _instance;
     return prefs?.clear() ?? Future.value(false);
   }
 }